pyproject.toml
[project]
name = "k8s-autopilot"
version = "0.1.0"
description = "Multi Agents responsible for complete Kubernetes deployment automation (generation → deployment → monitoring)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.10",
    "colorama>=0.4.6",
    "deepagents>=0.2.4",
    "langchain>=1.0.3",
    "langchain-core>=1.0.2",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-supervisor>=0.0.30",
    "langgraph-swarm>=0.0.2",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "python-toon>=0.1.3",
    "uvicorn>=0.38.0",
]

license = {text = "Apache-2.0"}
license-files = ["LICENSE"]
authors = [
    {name = "talkops"},
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://github.com/talkops-ai/k8s-autopilot"
repository = "https://github.com/talkops-ai/k8s-autopilot.git"
documentation = "https://github.com/talkops-ai/k-autopilot/blob/main/README.md"

[project.scripts]
k8s-autopilot = "k8s_autopilot.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.9.7",
    "mypy>=1.0.0",
    "ipykernel>=6.29.0",
    "rich>=14.2.0",
    "python-toon>=0.1.3",
    "asyncclick>=8.3.0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["k8s_autopilot"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "PLR0913", "PLR2004", "C901", "W191"]

[tool.ruff.lint.isort]
known-first-party = ["k8s_autopilot"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

